+++ src/main/java/chess/repository/ChessGameRepository.java
@@ -1,19 +1,15 @@
 package chess.repository;
 
-import static java.util.stream.Collectors.*;
-
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import chess.converter.StringToPositionConverter;
 import chess.domain.ChessGame;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 import chess.domain.state.GameState;
-import chess.repository.converter.BoardToStringConverter;
-import chess.repository.converter.StringToPieceConverter;
-import chess.repository.converter.StringToStateConverter;
+import chess.converter.BoardToStringConverter;
+import chess.converter.StringToStateConverter;
 import chess.repository.dao.ChessGameDao;
 import chess.repository.dao.TileDao;
 
@@ -40,17 +36,9 @@ public class ChessGameRepository implements GameRepository {
 			return Optional.empty();
 		}
 		Map<String, String> tiles = tileDao.selectByGameName(name);
-		GameState gameState = StringToStateConverter.of(state, convertToBoard(tiles));
+		GameState gameState = StringToStateConverter.of(state, tiles);
 
 		return Optional.of(new ChessGame(name, gameState));
-	}
-
-	private Map<Position, Piece> convertToBoard(Map<String, String> tiles) {
-		return tiles.entrySet().stream()
-			.collect(toMap(
-				entry -> StringToPositionConverter.from(entry.getKey()),
-				entry -> StringToPieceConverter.from(entry.getValue())
-			));
 	}
 
 	@Override
