+++ src/main/java/chess/webcontroller/WebController.java
@@ -13,8 +13,8 @@ import chess.domain.command.Command;
 import chess.domain.command.Start;
 import chess.domain.state.Ready;
 import chess.service.ChessGameService;
-import chess.webcontroller.converter.RequestToCommandConverter;
-import chess.webcontroller.converter.RequestToMapConverter;
+import chess.converter.RequestToCommandConverter;
+import chess.converter.RequestToMapConverter;
 import chess.webcontroller.dto.BoardResponseDto;
 import chess.webcontroller.dto.GameResponseDto;
 import chess.webcontroller.dto.NameResponseDto;
